Move encoding package to the top level
