+++ encoding/bytes.go
