+++ encoding/float.go
