+++ encoding/float_test.go
